Separate the GE "searchfile" command
